split file contents
