+++ libs/scalajslib/api/src/mill/scalajslib/worker/api/ESFeatures.scala
@@ -1,0 +1,8 @@
+package mill.scalajslib.worker.api
+
+private[scalajslib] case class ESFeatures(
+    allowBigIntsForLongs: Boolean,
+    avoidClasses: Boolean,
+    avoidLetsAndConsts: Boolean,
+    esVersion: ESVersion
+)
