+++ libs/scalajslib/api/src/mill/scalajslib/worker/api/ESModuleImportMapping.scala
@@ -1,0 +1,7 @@
+package mill.scalajslib.worker.api
+
+private[scalajslib] sealed trait ESModuleImportMapping
+
+private[scalajslib] object ESModuleImportMapping {
+  case class Prefix(prefix: String, replacement: String) extends ESModuleImportMapping
+}
