+++ libs/scalajslib/api/src/mill/scalajslib/worker/api/ESVersion.scala
@@ -1,0 +1,12 @@
+package mill.scalajslib.worker.api
+
+private[scalajslib] enum ESVersion {
+  case ES2015
+  case ES2016
+  case ES2017
+  case ES2018
+  case ES2019
+  case ES2020
+  case ES2021
+  case ES5_1
+}
