+++ libs/scalajslib/api/src/mill/scalajslib/worker/api/ModuleKind.scala
@@ -1,0 +1,7 @@
+package mill.scalajslib.worker.api
+
+private[scalajslib] enum ModuleKind {
+  case NoModule
+  case CommonJSModule
+  case ESModule
+}
