+++ libs/scalajslib/api/src/mill/scalajslib/worker/api/ModuleSplitStyle.scala
@@ -1,0 +1,9 @@
+package mill.scalajslib.worker.api
+
+private[scalajslib] sealed trait ModuleSplitStyle
+
+private[scalajslib] object ModuleSplitStyle {
+  case object FewestModules extends ModuleSplitStyle
+  case object SmallestModules extends ModuleSplitStyle
+  final case class SmallModulesFor(packages: List[String]) extends ModuleSplitStyle
+}
