+++ libs/scalajslib/api/src/mill/scalajslib/worker/api/OutputPatterns.scala
@@ -1,0 +1,9 @@
+package mill.scalajslib.worker.api
+
+private[scalajslib] final case class OutputPatterns(
+    jsFile: String,
+    sourceMapFile: String,
+    moduleName: String,
+    jsFileURI: String,
+    sourceMapURI: String
+)
